requirements.txt
# Diet AI — Python requirements
# Install with: pip install -r requirements.txt

# ── Core ML ───────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
accelerate>=0.27.0          # needed for device_map="auto"
sentencepiece>=0.1.99       # Mistral tokenizer dependency
protobuf>=3.20.0

# ── Image handling ─────────────────────────────────────────────
Pillow>=10.0.0

# ── Web UI ─────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Data & charts ──────────────────────────────────────────────
matplotlib>=3.8.0
pandas>=2.1.0
numpy>=1.26.0
seaborn>=0.13.0             # optional: for future heatmaps

# ── API / networking ───────────────────────────────────────────
requests>=2.31.0
